Small networking cleanup
Removes unused ping smoothing feature (didn't actually compile so safe
to say nobody is using it). It could cause problems for QC mods that use
ping for antilag purposes, and QC can easily average pings if desired.

Removes comment about never timing out loopback.
The check preventing loopback timeouts was removed when SV_CheckTimeouts
was added and IMO timeouts SHOULD apply on loopback. Someone can run a
server on their PC and connect to it and if their client fails or is
killed abruptly, it shouldn't occupy a slot forever.